backend/requirements.txt
arxiv>=2.1.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
